Renaming Example to ResponseExample.
